backend/autoHR/requirements.txt
asgiref==3.8.1
Django==5.0.4
djangorestframework==3.15.1
psycopg==3.1.18
psycopg-binary==3.1.18
PyMuPDF==1.24.1
PyMuPDFb==1.24.1
sqlparse==0.4.4
typing_extensions==4.11.0
